refactor: simplify provider config
